wip: created useAuth.ts within hooks directory
